Rename file to `scene-menu-items`
